+++ Cargo.toml
@@ -19,46 +19,15 @@ doc = false
 [lints]
 workspace = true
 
-[features]
-default = ["http", "jsondb", "messaging", "keyvalue", "p2p", "rpc", "vault"]
-http = ["dep:wasmtime-wasi-http"]
-jsondb = ["dep:jmespath", "dep:mongodb"]
-keyvalue = ["dep:async-nats"]
-messaging = ["dep:async-nats"]
-p2p = ["dep:iroh"]
-rpc = ["dep:async-nats"]
-vault = ["dep:azure_identity", "dep:azure_security_keyvault"]
-
 [dependencies]
-anyhow.workspace = true
-async-nats = { version = "0.40.0", optional = true }
-async-trait.workspace = true
-azure_identity = { version = "0.23.0", optional = true, default-features = false }
-azure_security_keyvault = { version = "0.21.0", optional = true, default-features = false }
-base64ct = { version = "1.7.3", features = ["std"] }
-bytes = "1.10.1"
+async-nats.workspace = true
 dotenv = "0.15.0"
-futures = "0.3.31"
-http = "1.3.1"
-hyper = "1.6.0"
-iroh = { version = "0.29.0", default-features = false, optional = true }
-jmespath = { version = "0.3.0", optional = true }
-mongodb = { version = "3.2.3", optional = true }
-nkeys = "0.4.4"
-runtime = { path = "crates/runtime" }
-serde.workspace = true
-serde_json.workspace = true
-sha2 = "0.10.9"
+runtime.workspace = true
+services = { path = "crates/services" }
 tokio.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
-url = "2.5.4"
 wasmtime.workspace = true
-wasmtime-wasi.workspace = true
-wasmtime-wasi-http = { workspace = true, optional = true }
-
-# ecdsa = "0.16.9"
-# k256 = "0.13.4"
 
 [profile.release]
 lto = true
@@ -100,8 +69,10 @@ pedantic = "warn"
 
 [workspace.dependencies]
 anyhow = "1.0.98"
+async-nats = "0.40.0"
 async-trait = "0.1.88"
 http = "1.3.1"
+runtime = { path = "crates/runtime" }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 tokio = { version = "1.45.0", features = ["signal"] }
