+++ crates/services/src/vault.rs
