+++ crates/services/wit/deps/keyvalue/atomic.wit
