+++ crates/services/wit/deps/keyvalue/batch.wit
