+++ crates/services/wit/deps/keyvalue/store.wit
