+++ crates/services/wit/deps/keyvalue/watch.wit
