+++ crates/services/wit/deps/keyvalue/world.wit
