+++ crates/services/wit/deps/messaging/consumer.wit
