+++ crates/services/wit/deps/messaging/guest.wit
