+++ crates/services/wit/deps/messaging/messaging.wit
