+++ crates/services/wit/deps/messaging/producer.wit
