+++ crates/services/wit/deps/messaging/types.wit
