+++ crates/services/wit/deps/p2p/container.wit
