+++ crates/services/wit/deps/p2p/world.wit
