+++ crates/services/wit/deps/vault/vault.wit
