+++ crates/services/wit/deps/vault/world.wit
